+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.8]
+### Fixed
+ - Generating link for `module` type
+ - `Return` block in methods
+ 
 ## [2.1.7]
 ### Fixed
  - Use Object.assign instead of spread in search